requirements.txt
#### BASE REQUIREMENTS ####

# The Python web framework on which this app is built
Django==4.0.6

# A REST API framework for Django projects
djangorestframework==3.13.1

# Django middleware which permits cross-domain API requests
django-cors-headers==3.13.0

# Library for writing reusable URL query filters
django-filter==22.1

# Simple JWT is a JSON Web Token authentication plugin for the Django REST Framework.
djangorestframework-simplejwt==5.2.0

# Python-dotenv reads key-value pairs from a .env file and can set them as environment variables. I
python-dotenv==0.20.0

#### _____ ####

# Generate real Swagger/OpenAPI 2.0 specifications from a Django Rest Framework API.
drf-yasg==1.20.0

# Model Bakery offers you a smart way to create fixtures for testing in Django.
model-bakery==1.6.0

# Faker is a Python package that generates fake data for you.
Faker==13.15.0

